fix unrecognized print format specifier for int8/uint8

+++ include/cute/util/print.hpp
@@ -70,13 +70,13 @@ print(char c) {
 CUTE_HOST_DEVICE
 void
 print(signed char a) {
-  printf("%hhd", a);
+  printf("%d", (int)(a));
 }
 
 CUTE_HOST_DEVICE
 void
 print(unsigned char a) {
-  printf("%hhu", a);
+  printf("%u", (unsigned int)(a));
 }
 
 CUTE_HOST_DEVICE
